+++ sale-service/src/main/java/vn/viettel/sale/entities/PoTransDetail.java
@@ -29,13 +29,13 @@ public class PoTransDetail extends BaseEntity {
     @Column(name = "QUANTITY")
     private Integer quantity;
     @Column(name = "PRICE")
-    private Float price;
+    private Double price;
     @Column(name = "PRICE_NOT_VAT")
-    private Float priceNotVat;
+    private Double priceNotVat;
     @Column(name = "AMOUNT")
-    private Float amount;
+    private Double amount;
     @Column(name = "AMOUNT_NOT_VAT")
-    private Float amountNotVat;
+    private Double amountNotVat;
     @Column(name = "RETURN_AMOUNT")
     private Integer returnAmount;
 }
